Worked on UI, Cinemachine
Got the health bar kinda working, tried adding a couple of lines of code to add health and that didn't work so I will need to try to fix that if I want to allow the player to gain health back.

Imported Cinemachine to try to make a smoother camera experience for the player. That didn't really work so I haven't actually added anything with Cinemachince to the scenes.

Worked on EnemyAI, Player chase still doesn't work. Need to find solution ASAP because for some reason it is still buggy af