Updated Player, added some prefabs
I had set up the player correctly in one scene and I had failed to set up the player correctly in the other. So I made a bunch of prefabs so that I could make testing enviroments very quickly in the project. So far this project has been going very well and I'm pleased with how far I've come in such a short amount of time.