Work on Laptop (9/24/18) (#1)
* EnemyAI mostly working on laptop

Got the enemy to chase the player, now I need to make the enemy be able to attack the player. Also made the Gizmos semi transparent so that you can actually see the player objects underneath the spheres.

* Added a function so that the enemy can attack the player

So far the EnemyAI is working very well. However, I want the function that checks to find the shortest distance between all of the points to run after the player exits the recognition zone. For some reason, "_nextPoint" is getting auto set to -1 and I don't know why...

* Updated Player, added some prefabs

I had set up the player correctly in one scene and I had failed to set up the player correctly in the other. So I made a bunch of prefabs so that I could make testing enviroments very quickly in the project. So far this project has been going very well and I'm pleased with how far I've come in such a short amount of time.