ENEMYAI COMPLETE!!!
I finally fixed the EnemyAI script. The problem was that the minDist variable ended up being too small and that there were no distances less than that number. So I just reset the minDist variable to infinity whenever the player is being chased. I spent over an hour last night trying to find the error and in 5 minutes this morning I actually found the error and fixed it :P