Diagonal Movement Added
One of the things that I've wanted to have in the game to set it apart from other games is two types of movement. When the player is in his standard enviroment, he will only be able to move up, down, left, and right. But when he enters the "dream" state, the player will also be able to move diagonally, adding depth to his movement and also making it mathematically possible for him to evade an enemy's pursuit. So far all of the scripts are working great and I think I'm ready to start making the story for this game.